// apps/web/src/api/__generated__/types/UpdateRecruitmentPostRequest.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * sync
 * OpenAPI spec version: 0.0.1
 */
import type { UpdateRecruitmentPostRequestContent } from './UpdateRecruitmentPostRequestContent';
import type { UpdateRecruitmentPostRequestEmploymentType } from './UpdateRecruitmentPostRequestEmploymentType';
import type { UpdateRecruitmentPostRequestExperienceLevel } from './UpdateRecruitmentPostRequestExperienceLevel';
import type { UpdateRecruitmentPostRequestRecruitmentStatus } from './UpdateRecruitmentPostRequestRecruitmentStatus';
import type { UpdateRecruitmentPostRequestReferencedPostIdsItem } from './UpdateRecruitmentPostRequestReferencedPostIdsItem';
import type { UpdateRecruitmentPostRequestTagsItem } from './UpdateRecruitmentPostRequestTagsItem';
import type { UpdateRecruitmentPostRequestWorkMode } from './UpdateRecruitmentPostRequestWorkMode';

export interface UpdateRecruitmentPostRequest {
  /**
   * 커버 이미지 미디어 ID
   * @nullable
   */
  coverMediaId?: string | null;
  /** 경력 수준 */
  experienceLevel: UpdateRecruitmentPostRequestExperienceLevel;
  /** 모집 상태 */
  recruitmentStatus: UpdateRecruitmentPostRequestRecruitmentStatus;
  /** 근무 방식 */
  workMode: UpdateRecruitmentPostRequestWorkMode;
  /** 고용 형태 */
  employmentType: UpdateRecruitmentPostRequestEmploymentType;
  /**
   * 기존 커버 삭제 여부
   * @nullable
   */
  removeCover?: boolean | null;
  /**
   * 근무 지역
   * @nullable
   */
  location?: string | null;
  /**
   * 참조 게시글 ID 목록
   * @nullable
   */
  referencedPostIds?:
    | UpdateRecruitmentPostRequestReferencedPostIdsItem[]
    | null;
  /** 모집 제목 */
  title: string;
  /**
   * 모집 마감 시각
   * @nullable
   */
  closesAt?: string | null;
  /** 모집 내용 */
  content: UpdateRecruitmentPostRequestContent;
  /**
   * 기술 태그 목록
   * @nullable
   */
  tags?: UpdateRecruitmentPostRequestTagsItem[] | null;
}